AndroidContentURI: Move code from the header to cpp. Some assorted cleanup, add a unit test for Download paths
